fourteenth program
A heap is a complete binary tree, and the binary tree is a tree in which the node can have the utmost two children. A complete binary tree is a binary tree in which all the levels except the last level, i.e., leaf node, should be completely filled, and all the nodes should be left-justified.

eap sort basically recursively performs two main operations -

Build a heap H, using the elements of array.
Repeatedly delete the root element of the heap formed in 1st phase.